Simple_Implementation/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
google-generativeai
duckduckgo-search
python-multipart
python-dotenv
